When IA finish, game end
