+++ Assets/Scripts/Race/CarLoader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bbc69ed716aad2741975867057127b76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
