+++ Assets/Scripts/Race/LastChekcpoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cd6674c44e9be4640aa0d889fb60009e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
