+++ Assets/Scripts/Race/LastChekpointIA.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 96bfac0e35f5e964d93f4db3c71b065c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
