--- Assets/Scripts/Timer/CarLoaderCity.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8f18b3b5a6f0bc644b8262d42aefafeb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
