+++ Assets/Scripts/Timer/TimerMode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2cc91fe0e60a0af48aa8f5b876d80396
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
